added testing and more UI fixes

+++ core/managers/user_manager.py
@@ -5,10 +5,16 @@ from collections import defaultdict
 from core.models.stock import Stock
 from core.models.user import User
 
+from core.managers.base_manager import Manager, ManagerException
 from core.managers.db_manager import DBManager
+from core.utils import password_utils
 
-class UserManager(object):
-	def __init__(self):
+class UserManagerException(ManagerException):
+	pass
+
+
+class UserManager(Manager):
+	def __init__(self, **kwargs):
 		self.db_manager = DBManager('users')
 
 	@classmethod
@@ -30,8 +36,8 @@ class UserManager(object):
 	def _serialize_user(cls, user):
 		if user:
 			return (user.get('id'), Time.get_time(), user.get('username'), user.get('first_name'), user.get('last_name'),
-			user.get('email'), user.get('password'), user.get('about'), user.get('location'), user.get('website'),
-			user.get('image_url'), user.get('gender'), user.get('birthday'), '')
+			user.get('email'), password_utils.encrypt_password(user.get('password')), user.get('about'), user.get('location'), user.get('website'),
+			user.get('image_url'), user.get('gender'), user.get('birthday'), user.get('stock_list'))
 
 		return ()
 
@@ -41,17 +47,27 @@ class UserManager(object):
 		
 		return self.db_manager.add_one(query, UserManager._serialize_user(user))
 
-	def update_one(self, user_id, **kwargs):
-		query =  " "
+	def update_one(self, user):
+		if user:
+			query =  " "
 
-		for key, value in kwargs.items():
-			query += "%s = \"%s\", " % (key, value)
+			serialized_user = UserManager._serialize_user(user)
 
-		query = query[:-2]
+			count = 0
 
-		query =  query + ''' WHERE user_id=%s''' % user_id
+			for value in serialized_user:
+				if User.DB_MAPPINGS[count] not in User.unmutable_fields():
+					query += "%s = \"%s\", " % (User.DB_MAPPINGS[count], value)
+				
+				count = count + 1
 
-		return self.db_manager.update_one(query)
+			query = query[:-2]
+
+			query =  query + ''' WHERE user_id=%s''' % user.get('id')
+
+			return self.db_manager.update_one(query)
+
+		raise UserManagerException('UserIsNoneError')
 
 	def get_one(self, **kwargs):
 		query =  ""
@@ -65,13 +81,24 @@ class UserManager(object):
 
 		return UserManager._deserialize_user(user)
 
-	def delete_one(self, **kwargs):
-		query =  ""
+	def delete_one(self, user):
+		if user:
+			query = "user_id = \"%s\"" % user.get('id')
 
-		for key, value in kwargs.items():
-			query += "%s = \"%s\" AND " % (key, value)
+			return self.db_manager.delete_one(query)
 
-		query = query[:-5]
+		raise UserManagerException('UserIsNoneError')
 
-		return self.db_manager.delete_one(query)
+	def login(self, username_or_email, password):
+		if "@" in username_or_email:
+			user = self.get_one(email=username_or_email)
+		else:
+			user = self.get_one(username=username_or_email)
 
+		if not user:
+			return None
+
+		if not user.check_password(password):
+			return None
+
+		return user
